Read personalis Small Variant reports for a single sample.

We only read in the "Preferred Transcript" report here:

read_personalis_small_variant_report_sample(
  sample_folder,
  modality,
  sample_type
)

Arguments

sample_folder

path to personalis folder

modality

modality from which the variants were called. Can be either 'DNA' or 'RNA'

sample_types

Can be ome or multiple of of 'tumor', 'normal', or 'somatic'. 'tumor' refers to tumor sample vs. genome reference (i.e. somatic+germline mutations), 'normal' refers to normal sample vs. genome reference (i.e. germline mutations) and 'somatic' refers to tumor vs. normal (i.e. somatic mutations only).

Value

data frame with all variants

Details

Preferred Transcript: The RefSeq accession.version for the transcript used for variant analysis. Personalis uses a curated list of transcripts, which is based on the number of times a transcript (accession.version) is referred to in COSMIC. If not present in COSMIC, the default transcript would be the one corresponding to the longest CDS. (source: Personalis Analysis_Pipeline_Documentation)

We also do not read the cancer_clinical_research, the cancer_research and the lowpupulationfreq reports, because they are subsets of the full report.

In addition, Personalis also provides raw VCF files with all (unfiltered) variants. We currently don't read them in because without additional filtering they are not very useful. If you need this level of information, feel free to start from the raw vcf files or even run your own variant calling based on the FASTQ files.