Ensure that every col in col data has an entry in the bumpy matrix

When splitting data frames into a bumpy matrix, samples that don't have a row in the data frame, won't result in a column of the bumpy matrix. This is problematic, as it a sample might have been successfully measured, but simply not resulted in a call of a CNV, mutation, etc.

add_dummy_entry(df, col_data, sample_col = "sample")

Arguments

df

data.frame data frame that will be converted into a BumpyMatrix

col_data

data.frame data frame that is used as colData (must have rownames that are sample identifiers!)

sample_col

character column in df that contains the sample identifier

Value

tibble new data frame with dummy entries added

Details

Loosing the sample would result in loss of information:

we loose the information that there is "no mutation"

there might be metadata in colData that would be lost

While this is not an ideal solution, this function provides a workaround by adding rows that are all NAs for the missing samples. When converting the bumpy matrix back to a data frame, the dummy entries are lost, so we end up with the original data frame, which is nice!.